allure-python/allure-python-pytest/requirements.txt
pytest>=5.0.0
pytest-rerunfailures>=9.1.1
allure-pytest>=2.8.18
flake8>=3.8.3